+++ tests/unit/cli/test_add_command_complete.py
@@ -19,7 +19,7 @@ from datacompose.cli.commands.add import (  # noqa: E402
     _run_add,
 )
 from datacompose.cli.main import cli  # noqa: E402
-from datacompose.transformers.discovery import TransformerDiscovery  # noqa: E402
+from datacompose.cli.discovery import TransformerDiscovery  # noqa: E402
 
 
 @pytest.mark.unit
