+++ tests/unit/transformers/test_discovery.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from datacompose.transformers.discovery import TransformerDiscovery
+from datacompose.cli.discovery import TransformerDiscovery
 
 # Add project root to path for imports
 project_root = Path(__file__).parent.parent.parent.parent
